% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{addQcColsToMeta}
\alias{addQcColsToMeta}
\title{add qc metrics to metadata}
\usage{
addQcColsToMeta(meta_df, qc_df)
}
\arguments{
\item{meta_df}{metadata for the samples you'd like to QC. note that these
must be included in the pipeline_output_dirpath}

\item{qc_df}{output of the function \link{novoalignPipelineQC}}
}
\value{
a subset of the meta_df columns suitable for auto/manual auditing
}
\description{
take the long form of the qc_df and add the appropriate metrics
to the metadata in a 'wide' format
}
