% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLogExpression.R
\name{addRleSummaryStats}
\alias{addRleSummaryStats}
\title{calculate RLE by replicate groups}
\usage{
addRleSummaryStats(object)
}
\arguments{
\item{object}{a brentlabRnaSeqSet with colData replicate_group}
}
\value{
a list of dataframes for each replicate group in replicateS_sample_list, each with dimensions gene x sample. values are RLE of the gene in a given sample
}
\description{
calculate RLE by replicate groups
}
