% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{addTablesToQcDatabase}
\alias{addTablesToQcDatabase}
\title{add new_qc_database tables to old_qc_database_tables}
\usage{
addTablesToQcDatabase(
  old_qc_database_path,
  new_qc_database_path,
  qc_tablenames
)
}
\arguments{
\item{old_qc_database_path}{connection to existing qc database -- the
one to be added to}

\item{new_qc_database_path}{connection to the new qc database}

\item{qc_tablenames}{the tables to be added}
}
\description{
add new_qc_database tables to old_qc_database_tables
}
