% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{autoAuditQcTable}
\alias{autoAuditQcTable}
\title{Audit a qc table with metrics added}
\usage{
autoAuditQcTable(qc_table)
}
\arguments{
\item{qc_table}{output of \link{addQcColsToMeta}}
}
\value{
the qc_table with autoStatus and autoStatusDecomp columns added
}
\description{
audits the output of \link{addQcColsToMeta}
}
