% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{brentlabRnaSeqExperiment-class}
\alias{brentlabRnaSeqExperiment-class}
\alias{brentlabRnaSeqExperiment}
\title{brentlabRnaSeqExperiment constructor.}
\usage{
brentlabRnaSeqExperiment(brentlabRnaSeqSet, set_name = "")
}
\arguments{
\item{brentlabRnaSeqSet}{A brentlabRnaSeqSet}

\item{set_name}{one of {'ninetyMin_2016Grant', 'ninetyMin_2016GrantWithDoubles',
'ninetyMin_non2016grant', 'ninetyMin_all',  'envPert_epWT',
'envPert_perturbed', 'envPert_titrationWT'}}
}
\value{
a brentlabRnaSeqExperiment object
}
\description{
Extends brentlabRnaSeqSet to use with pre-specified experiment
sets, eg Environmental Perturbation. Intended for lower level use, primarily.
}
