% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{brentlabRnaSeqSet-class}
\alias{brentlabRnaSeqSet-class}
\alias{brentlabRnaSeqSet}
\alias{brentlabRnaSeqSetFromDatabase}
\title{brentlabRnaSeqSet object and constructor}
\usage{
brentlabRnaSeqSet(dds = NULL)

brentlabRnaSeqSetFromDatabase(
  organism,
  username,
  password,
  full_rnaseq_only = TRUE
)
}
\arguments{
\item{dds}{a deseq data object. Passing a null object will instantiate with
the \link[DESeq2:makeExampleDESeqDataSet]{DESeq2::makeExampleDESeqDataSet}}

\item{organism}{currently, only kn99}

\item{username}{db username}

\item{password}{db password}

\item{full_rnaseq_only}{filter out spikeins before returning. Default TRUE}
}
\value{
a brentlabRnaSeqSet for the given organism
}
\description{
A container for expression data in brentlab rnaseq databases
}
\seealso{
\itemize{
\item \href{https://github.com/mikelove/DESeq2}{extending DESeqDataSet}
\item \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/Extensions.html}{extending SummarizedExperiment}.
Note that A \link{SummarizedExperiment} object becomes a \link{RangedSummarizedExperiment}
object when rowRanges are added.
}
}
