% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{brentlabRnaSeqSetTransform-class}
\alias{brentlabRnaSeqSetTransform-class}
\alias{brentlabRnaSeqSetTransform}
\title{brentlabRnaSeqSetTransform object and constructor. This extends DESeqTransform.}
\usage{
brentlabRnaSeqSetTransform(SummarizedExperiment)
}
\arguments{
\item{SummarizedExperiment}{a RangedSummarizedExperiment}
}
\value{
a brentlabRnaSeqSetTransform object
}
\description{
Extends DESeqTransform for use in rleTransform
It is used by \code{\link{rleTransform}}
}
