% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{calculate coverage over a given locus}
\usage{
calculateCoverage(
  bamfile_path,
  annote_db,
  gene_id,
  strandedness,
  align_expr_prefix = Sys.getenv("ALIGN_EXPR_PREFIX"),
  bam_suffix = "_sorted_aligned_reads_with_annote.bam"
)
}
\arguments{
\item{bamfile_path}{path to a bam file @seealso brentlabRnaSeqTools::createBamPath()}

\item{annote_db}{a GenomicFeatures TxDb object. Maybe one made from a gtf, eg txdb = makeTxDbFromGFF("data/liftoff_h99_to_kn99.gtf", format = "gtf")}

\item{gene_id}{a gene_id of interest -- must be in the gene names of the annote_db object}

\item{strandedness}{one of c("unstranded", "reverse") indicating the strandedness of the library. note: forward not currently supported}

\item{align_expr_prefix}{= path to the directory which stores the run_12345_samples run directories.
For example, /lts/mblab/Crypto/rnaseq_data/lts_align_expr.
By default, this looks in your .Renviron for a key ALIGN_EXPR_PREFIX}

\item{bamfile_suffix}{= whatever is appended after the fastqFileName (no extension).
Default is "_sorted_aligned_reads_with_annote.bam"}
}
\value{
percent coverage of feature with reads above a given quality threshold and coverage depth threshold (see getCoverageOverRegion())
}
\description{
calculate coverage over a given locus
}
\seealso{
\code{\link{getCoverageOverRegion}}, \code{\link{strandedScanBamParam}}
}
