% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{calculateGeneWiseMedians}
\alias{calculateGeneWiseMedians}
\title{calculate medians across rows of dataframe}
\usage{
calculateGeneWiseMedians(count_df)
}
\arguments{
\item{count_df}{could be any numeric dataframe, but in this context it will typically be a count (raw or log2) df}
}
\value{
a vector of row-wise medians (length == nrow of input df)
}
\description{
calculate medians across rows of dataframe
}
