% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{coerceToDds}
\alias{coerceToDds}
\alias{coerceToDds,brentlabRnaSeqSet-method}
\title{coerce brentlabRnaSeqSet to DESeqDataSet}
\usage{
coerceToDds(x, ...)

\S4method{coerceToDds}{brentlabRnaSeqSet}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqSet}

\item{...}{other arguments}
}
\description{
coerce brentlabRnaSeqSet to DESeqDataSet
}
