% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{compileQualimapSummaryMetrics}
\alias{compileQualimapSummaryMetrics}
\title{parse qualimap summary output}
\usage{
compileQualimapSummaryMetrics(sample_name, qualimap_results)
}
\arguments{
\item{sample_name}{name of the sample which will go into the Sample column.
See example}

\item{qualimap_results}{path to rnaseq_qc_results.txt for a given sample}
}
\value{
a dataframe with columns sample, exonic, intergenic, overlapping_exon,
count, percent
}
\description{
extract metrics from the file rnaseq_qc_results.txt in nf-co/rnaseq qualimap
output
}
\examples{
\dontrun{
   pool_plate2_qualimaps_files = Sys.glob("path/to/qualimap/*/rnaseq_qc_results.txt")
   names(pool_plate2_qualimaps_files) =
       basename(dirname(pool_plate2_qualimaps_files))
   pool_plate2_qualimap_df = bind_rows(lapply(names(pool_plate2_qualimaps_files),
       function(x)
           compileQualimapSummaryMetrics(x, pool_plate2_qualimaps_files[[x]]))) \%>\%
           arrange(Sample)
}

}
