% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{compileTinOutput}
\alias{compileTinOutput}
\title{extract all metrics from the output of tin.py run on a batch of files}
\usage{
compileTinOutput(tin_output_dir, bam_suffix = ".markdup.sorted.bam")
}
\arguments{
\item{tin_output_dir}{directory containing the rnaseqc output}

\item{bam_suffix}{the suffix to remove from the bam file sample names. Default to '.markdup.sorted.bam'
for nf-co/rnaseq_pipeline star_salmon output \url{https://nf-co.re/rnaseq}}
}
\value{
a list with items full_table, subset, rRna_vs_epe, rRna_vs_epe (see description)
}
\description{
presumably tin.py was run on a batch of samples, the output of
which is in some output dir. This function compiles all the
summary sheets and compiles them into a single dataframe
}
\note{
Estimated Library Complexity is a Picard metric and is similar to markDuplicates and totalDeduplicatedPercent
\url{https://gatk.broadinstitute.org/hc/en-us/articles/360037591931-EstimateLibraryComplexity-Picard-}
}
\seealso{
\url{http://rseqc.sourceforge.net/#tin-py}
}
