% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{connectToDatabase}
\alias{connectToDatabase}
\title{Connect to a remote postgresql database}
\source{
\url{https://rpostgres.r-dbi.org/}
}
\usage{
connectToDatabase(
  database_host,
  database_name,
  database_user,
  database_password
)
}
\arguments{
\item{database_host}{if connecting to a database hosted on AWS, it might be something like ec2-54-83-201-96.compute-1.amazonaws.com}

\item{database_name}{name of the database, eg for cryptococcus kn99, the database might be named kn99_database. Check with the documentation, whoever set up the database, or get into the server and check directly}

\item{database_user}{a user of the actual database, with some level of permissions. You'll need to check with the database maintainer for this. It is suggested that you use a .Renviron file in your local project (make sure it is completely ignored by git, R, etc) to store this info}

\item{database_password}{password to the database user. You'll need to check with the database maintainer for this. It is suggested that you use a .Renviron file in your local project (make sure it is completely ignored by git, R, etc) to store this info}
}
\value{
A DBI connection to the remote database
}
\description{
Use the RPostgres package to connect to a remote postgresql database
}
\note{
for information on using R environmental files, see \url{https://support.rstudio.com/hc/en-us/articles/360047157094-Managing-R-with-Rprofile-Renviron-Rprofile-site-Renviron-site-rsession-conf-and-repos-conf}
}
