% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{countLibrary}
\alias{countLibrary}
\title{Count reads in a library like HTSeq}
\usage{
countLibrary(
  bam_path,
  annote_db,
  feature_type,
  strandedness,
  num_threads = 1,
  single_end_reads_flag = TRUE,
  mapq_filter = 20L
)
}
\arguments{
\item{bam_path}{path to a bam file. the index must also exist}

\item{annote_db}{txdb object. see \code{\link[GenomicFeatures]{makeTxDbFromGFF}}}

\item{feature_type}{one of c("exon", "cds"). Determines what feature to count over}

\item{strandedness}{strandedness of library determined by protocol. One of c("reverse", "same", "unstranded")
reverse means the counted reads will be on the opposite strand of the feature, same means they
will be on the same strand. unstranded counts reads over the feature regardless of strand}

\item{num_threads}{number of threads available for parallelization. Default to 1}

\item{single_end_reads_flag}{whether or not the library is single or paired. Note: this is not written to deal
with paired end reads currently, though you can try}

\item{mapq_filter}{quality filter on read alignment. Default to 20, same as HTSeq.
see \url{https://htseq.readthedocs.io/en/release_0.11.1/count.html}}
}
\value{
A RangedSummarizedExperiment. See \code{\link[GenomicAlignments]{summarizeOverlaps}}
}
\description{
This function creates a granges object from the txdb and calls \code{\link{countReadsInRanges}}.
Uses for this function may be during QC to generate log2cpm over the CDS regions in protein
coding transcriptome, for example
}
\examples{
\dontrun{library(brentlabRnaSeqTools)
library(AnnotationDbi)
library(tidyverse)

kn99_db = loadDb("data/kn99_db.sqlite")

combined_df = getMetadata(database_info$kn99_host,
                          database_info$kn99_db_name,
                          Sys.getenv("db_username"),
                          Sys.getenv("password"))

run_5102 = combined_df \%>\%
 filter(runNumber == 5102,
        experimentDesign == '90minuteInduction',
        genotype1 == "CNAG_00000")

wt_fastq = run_5102[[1, 'fastqFileName']]

bam_path = createBamPath(5102,
                         wt_fastq,
                         align_expr_prefix =
                         "/mnt/htcf_scratch/chasem/rnaseq_pipeline/align_count_results")

cds_count = countLibrary(bam_path, kn99_db, 'cds', 'reverse', 8)

cds_counts_df = as_tibble(assay(cds_counts), rownames = "gene_name")}

}
\seealso{
\url{http://achri.blogspot.com/2016/02/how-not-to-use-deseq2-for-illumina.html},
\url{https://htseq.readthedocs.io/en/release_0.11.1/count.html},
\code{\link[GenomicAlignments]{summarizeOverlaps}}
}
