% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{countReadsInRanges}
\alias{countReadsInRanges}
\title{given a bam file path, GRanges object, and strandedness of library, return total counts}
\usage{
countReadsInRanges(bamfile_path, granges_of_interest, strandedness)
}
\arguments{
\item{bamfile_path}{path to bam file}

\item{granges_of_interest}{a GRanges object}

\item{strandedness}{one of c("reverse", "unstranded")}
}
\description{
given a bam file path, GRanges object, and strandedness of library, return total counts
}
