% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingCardsFunctions.R
\name{create_partitions}
\alias{create_partitions}
\title{Given a vector length adn the size of each parition, create a vector which
represents those partitions.}
\usage{
create_partitions(vector_length, equal_parts = 100)
}
\arguments{
\item{vector_length}{the total lenght of the partition vector}

\item{equal_parts}{how large should each partition be?}
}
\value{
a vector of \code{vector_length} divded into \code{equal_parts} with possibly
one additional vector of size less than \code{equal_parts}.
}
\description{
There may be one more partition of size less than the equally
divided parts. eg if the vector length is 14 and the desired partitions are
size 3, then there will be 4 partitions of length 3 and one of length 2,
1 1 1 2 2 2 3 3 3 4 4 4 5 5.
}
