% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{createDatabase}
\alias{createDatabase}
\title{create the sqlite database as input to qcExplorer}
\usage{
createDatabase(sample_df, status_df, qc_df_list, database_output, id_column)
}
\arguments{
\item{sample_df}{a tibble with an id column which corresponds to the id column
of the QC metrics tables and columns containing metadata on samples}

\item{status_df}{a tibble with columns metric, alias, threshold, status,
comparative for use in flagging samples}

\item{qc_df_list}{list of qc table dataframes -- already read in.
see articles in documentation
reports from the nf-co/rnaseq pipeline}

\item{database_output}{output path (full, with .sqlite extension) for the
sqlite database}

\item{id_column}{common id column between sample_df and QC tables}
}
\description{
sample_df (metadata on the samples) and any QC tables must
share a common ID column. qc_metrics_df contains the column names of the QC
metrics with columns threshold, status and comparative which store information
on how to flag samples which violate the threshold. Writes database to
specified file
}
