% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createEnvPert_epWT}
\alias{createEnvPert_epWT}
\title{filter combined_df for environmental perturbation sample set}
\usage{
createEnvPert_epWT(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
environmental pertubation set
}
\description{
filter combined_df for environmental perturbation sample set
}
