% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalSetFunctions.R
\name{createEnvPertSet}
\alias{createEnvPertSet}
\title{filter combined_df for environmental perturbation sample set}
\usage{
createEnvPertSet(combined_df)
}
\arguments{
\item{combined_df}{the combined tables of the database, returned directly from getMetadata() (meaning, the df hasn't been augmented after pulling from the database)}
}
\value{
environmental pertubation set
}
\description{
filter combined_df for environmental perturbation sample set
}
