% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createEnvPertSet_perturbed}
\alias{createEnvPertSet_perturbed}
\title{create Environmental Perturbation Perturbed set}
\usage{
createEnvPertSet_perturbed(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
environmental pertubation brentlabRnaSeqSet
}
\description{
create Environmental Perturbation Perturbed set
}
