% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createEnvPertSet_titrationWT}
\alias{createEnvPertSet_titrationWT}
\title{create Environmental Perturbation Titration (TDY451 only) set}
\usage{
createEnvPertSet_titrationWT(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
environmental pertubation set
}
\description{
create Environmental Perturbation Titration (TDY451 only) set
}
