% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{createExperimentSet}
\alias{createExperimentSet}
\alias{createExperimentSet,brentlabRnaSeqSet-method}
\title{Create a pre-defined experiment set}
\usage{
createExperimentSet(x, ...)

\S4method{createExperimentSet}{brentlabRnaSeqSet}(x, set_name)
}
\arguments{
\item{x}{a brentlabRnaSeqSet object}

\item{...}{other arguments}

\item{set_name}{one of {'ninetyMin_2016Grant', 'ninetyMin_2016GrantWithDoubles',
'ninetyMin_non2016grant', 'ninetyMin_all',  'envPert_epWT',
'envPert_perturbed', 'envPert_titrationWT'}}
}
\description{
for pre-defined experiment definitions, see the github repo
R/ExperimentSetFunctions.R
}
