% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{createFullRLETable}
\alias{createFullRLETable}
\title{calculate RLE}
\usage{
createFullRLETable(counts_df, gene_id_column = NULL, logged = FALSE)
}
\arguments{
\item{counts_df}{gene by samples dataframe of raw counts or logged counts (see paramter logged)}

\item{gene_id_column}{(Optional) if gene_id_column is not passed, a sequence from 1 to nrow(counts_df) will be used}

\item{logged}{(Default FALSE) set to true if log2 transformed counts are passed}
}
\value{
rle dataframe with genes x rle statistics
}
\description{
You may pass raw counts or log2 transformed counts with the logged = TRUE argument
}
