% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgvFunctions.R
\name{createIgvBatchscript}
\alias{createIgvBatchscript}
\title{create IGV viewer batch script (single range, as many tracks as you like)}
\usage{
createIgvBatchscript(
  bam_list,
  granges,
  igv_genome,
  output_dir,
  output_file_basename,
  maxPanelHeight = 500
)
}
\arguments{
\item{bam_list}{list of bamfiles -- multiple files means multiple tracks}

\item{granges}{granges describing the range you wish to visualize -- range
must be on one chromosome/contig}

\item{igv_genome}{.genome file created by IGV tools}

\item{output_dir}{where to deposit both the script and the browser shots}

\item{output_file_basename}{this will serve as both the name of the browser
shot after running IGV, and the name o the batchscript itself}

\item{maxPanelHeight}{default 500. height of the IGV window}
}
\value{
None. This prints a batch script to the output dir
}
\description{
this will create the following:
new                                # batchscript keyword new (new snapshot)
snapshotDirectory IGV_Snapshots    # output directory
maxPanelHeight 500                 # maximum height of igv browser viewer
genome organism.genome             # igv .genome file
load control.bam                   # load alignment files in bam_file_list
load perturbed.bam
goto chr1:1-100                    # load region of interest
snapshot batchfilename_locus1.png  # saves a snapshot of the IGV window to an image file
goto chr10:3-300                   # repeat at another locus
snapshot batchfilename_locus2.png
exit                               # quit session
}
\examples{
\dontrun{
# You need to have igv installed on the machine where you will run this. This
# is something you can run on your local. Create the browser shot like so:

xvfb-run --auto-servernum igv.sh -b script.bat

}

}
