% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallyFunctions.R
\name{createInductionSetTally}
\alias{createInductionSetTally}
\title{create 90 minute induction set tally}
\usage{
createInductionSetTally(
  induction_meta_qual,
  sorted_passing_induction_meta_qual,
  iqr_fltr_rle_summary,
  grant_df
)
}
\arguments{
\item{induction_meta_qual}{the metadata of the entire set, unfiltered}

\item{sorted_passing_induction_meta_qual}{metadata (with quality columns) filtered for manual/auto status}

\item{iqr_fltr_rle_summary}{sorted_passing_meta_qual filtered for IQR}

\item{grant_df}{the definition of the 90minuteInduction set. This object is available in the brentlabRnaSeqTools
package}
}
\description{
create 90 minute induction set tally
}
