% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLibraryProtocolLibraryDateModelMatrix.R
\name{createLibrarydateModelMatrix}
\alias{createLibrarydateModelMatrix}
\title{Create the libraryProtocol + libraryDate model matrix with the earliest date of each library protocol dropped}
\usage{
createLibrarydateModelMatrix(metadata_df)
}
\arguments{
\item{metadata_df}{the joined tables of the database (biosample to quality assess)}
}
\value{
a model matrix constructed as specified in the description
}
\description{
Create the libraryProtocol + libraryDate model matrix with the earliest date of each library protocol dropped
}
