% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{createMergedTable}
\alias{createMergedTable}
\title{Join QC tables (not qc_metrics) in qcExplorer input sqlite database}
\usage{
createMergedTable(
  ui_selected_columns,
  id_field,
  sample_df,
  status_df,
  qc_database_path
)
}
\arguments{
\item{ui_selected_columns}{a character vector of the selected columns}

\item{id_field}{the common field between all QC tables}

\item{sample_df}{name of the table containing sample metadata}

\item{status_df}{name of the table containing the QC metrics (this table
does not have the id column)}

\item{qc_database_path}{path to the sqlite qc database}
}
\value{
the merged database based input
}
\description{
used to create the table displayed in the qcExplorer
}
