% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_pipeline.R
\name{createNfCorePipelineSampleSheet}
\alias{createNfCorePipelineSampleSheet}
\title{create nf-co sample sheet}
\usage{
createNfCorePipelineSampleSheet(
  metadata,
  sample_columns,
  sequence_dir_prefix,
  check_files_flag = FALSE
)
}
\arguments{
\item{metadata}{most likely from the database. must include runNumber, fastqFileName, libraryProtocol,
and the columns used to create the sample name (see param sample_columns)}

\item{sample_columns}{a character vector of columns to concat into a sample identifier.
eg c("genotype1", "treatment", "timePoint", "floodmedia", "inductionDelay")}

\item{sequence_dir_prefix}{path to directory where the sequence run directories are stored.
eg /mnt/htcf_lts/lts_sequence}

\item{check_files_flag}{default FALSE. Set to true to check that files exist}
}
\value{
a dataframe with the columns which correspond to those required for the nf-co rnaseq pipeline
sample sheet. See \url{https://nf-co.re/rnaseq/3.2/usage#samplesheet-input}
}
\description{
create a samplesheet for the nf-co rnaseq pipeline
\url{https://nf-co.re/rnaseq/3.2/usage#samplesheet-input}
}
\note{
: currently hard coded to set strandedness to 'reverse' if libraryProtocol is E7420L, or 'unstranded',
AND for single strand libraries (fastq2 is an empty string)
}
