% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createNinetyMinuteInduction_2016grant}
\alias{createNinetyMinuteInduction_2016grant}
\title{90minuteInduction set Definition}
\usage{
createNinetyMinuteInduction_2016grant(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
the set metadata -- single KO only
}
\description{
The current definition of the 90 minute induction dataset,
according to the 2016 grant summary (loaded into environment, see head(grant_df)) -- single KO only
}
