% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createNinetyMinuteInduction_2016grantWithDoubles}
\alias{createNinetyMinuteInduction_2016grantWithDoubles}
\title{ninetyMin set where both doubles are in the grant_df}
\usage{
createNinetyMinuteInduction_2016grantWithDoubles(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
the set metadata
}
\description{
ninetyMin set where both doubles are in the grant_df
}
