% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{createNinetyMinuteInduction_non2016grant}
\alias{createNinetyMinuteInduction_non2016grant}
\title{Samples with genotypes not in the grant_df}
\usage{
createNinetyMinuteInduction_non2016grant(blrs)
}
\arguments{
\item{blrs}{a brentlabRnaSeqSet object}
}
\value{
the set metadata
}
\description{
return a set with only those genotypes (including doubles) not in the grant_df.
}
