% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNinetyMinuteInductionSet.R
\name{createNinetyMinuteInductionSet}
\alias{createNinetyMinuteInductionSet}
\title{The current definition of the 90 minute induction dataset, according to the 2016 grant summary (loaded into environment, see head(grant_df)) -- single KO only}
\usage{
createNinetyMinuteInductionSet(metadata, grant_df)
}
\arguments{
\item{metadata}{is the combined tables of the metadata database}

\item{grant_df}{is the 2016 grant summary TODO: put this in DATA}
}
\value{
the set metadata -- single KO only
}
\description{
The current definition of the 90 minute induction dataset, according to the 2016 grant summary (loaded into environment, see head(grant_df)) -- single KO only
}
