% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalSetFunctions.R
\name{createNinetyMinuteInductionWithDoubles}
\alias{createNinetyMinuteInductionWithDoubles}
\title{The current definition of the 90 minute induction dataset, according to the 2016 grant summary (loaded into environment, see head(grant_df)) -- single and double KO}
\usage{
createNinetyMinuteInductionWithDoubles(metadata, grant_df)
}
\arguments{
\item{metadata}{is the combined tables of the metadata database}

\item{grant_df}{is the 2016 grant summary TODO: put this in DATA}
}
\value{
the set metadata
}
\description{
The current definition of the 90 minute induction dataset, according to the 2016 grant summary (loaded into environment, see head(grant_df)) -- single and double KO
}
