% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextflowProcessing.R
\name{createNovoalignPipelineSamplesheet}
\alias{createNovoalignPipelineSamplesheet}
\title{create a sample sheet for the novoalign pipeline}
\usage{
createNovoalignPipelineSamplesheet(meta_df, sequence_dir_prefix)
}
\arguments{
\item{meta_df}{metadata for the samples you'd like to QC. note that these
must be included in the pipeline_output_dirpath}

\item{sequence_dir_prefix}{filepath to append to the run_1234_samples/fastq.fq.gz
eg "/scratch/mblab/chasem/rnaseq_pipeline/scratch_sequence"}
}
\value{
a dataframe in correct format for novo+hsteq pipeline
}
\description{
create a sample sheet for the novoalign pipeline
}
