% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{createQCdatabase}
\alias{createQCdatabase}
\title{create a sqlite database to hold the 'custom' qc data}
\usage{
createQCdatabase(database_dirpath)
}
\arguments{
\item{database_dirpath}{path to containing directory of new qc database}
}
\value{
database path
}
\description{
create a sqlite database to hold the 'custom' qc data
}
