% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{createStatusDf}
\alias{createStatusDf}
\title{create the metric status dataframe used to set thresholds and flag samples
in the QC explorer}
\usage{
createStatusDf(user_metric_vector, multiqc_output)
}
\arguments{
\item{user_metric_vector}{either an empty string if there are no user
defined metrics, or a character vector of metrics, eg c("epe", "perturbed_coverage")}

\item{multiqc_output}{path to the multiqc output data directory of
nf-co/rnaseq_pipeline}
}
\value{
a data frame with columns source, metric, alias, threshold, status
}
\description{
columns should be source, which corresponds to the tablenames
in the QC explorer database, metric, which corresponds the field in the table
threshold, which may be set by the user and is initially NA,
comparative
}
