% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{createTestTrainSet}
\alias{createTestTrainSet}
\alias{createTestTrainSet,brentlabRnaSeqExperiment-method}
\title{create test train set}
\usage{
createTestTrainSet(x, ...)

\S4method{createTestTrainSet}{brentlabRnaSeqExperiment}(x, min_set_size)
}
\arguments{
\item{x}{a brentlabRnaSeqExperiment object}

\item{...}{other arguments}

\item{min_set_size}{the minimum replicate set size to consider for hold outs}
}
\value{
a list with slots train and test, each with brentlabRnaSeqSetExperiments
}
\description{
For brentlabRnaSeqExperiment objects, this offers methods to
create train test sets
}
