% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\docType{data}
\name{database_colnames}
\alias{database_colnames}
\title{set global variables, mostly as a hack to make the tidyverse data binding pass the R CMD Check}
\format{
An object of class \code{character} of length 118.
}
\usage{
database_colnames
}
\description{
set global variables, mostly as a hack to make the tidyverse data binding pass the R CMD Check
}
\keyword{datasets}
