% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{database_info}
\alias{database_info}
\title{URLS to active databases}
\format{
A list of lists, organized by organism. Currently contains info for
KN99 and S288C_R64 databases
\describe{
\item{host}{host of the database server, eg ec2-18-224-181-136.us-east-2.compute.amazonaws.com}
\item{db_name}{cryptococcus database name. something like kn99_database}
\item{urls}{urls to all tables in database}
...
}
}
\source{
\url{https://rnaseq-databases-documentation.readthedocs.io/en/latest/}
}
\usage{
database_info
}
\description{
A list containing the urls to active databases. Named by organism (eg 'kn99' or 's288cr64')
}
\keyword{datasets}
