% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{database_urls}
\alias{database_urls}
\title{URLS to the yeast and kn99 databases}
\format{
An object of class \code{list} of length 2.
}
\usage{
database_urls
}
\description{
this will eventually hold the urls to all the tables for yeast and kn99
}
\keyword{datasets}
