% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{decomposeStatus2Bit}
\alias{decomposeStatus2Bit}
\title{decompose sums of powers of two to a list of the summed powers}
\usage{
decomposeStatus2Bit(status)
}
\arguments{
\item{status}{an integer that represents the sum of powers of 2}
}
\value{
a string of powers of 2 representing the bit, eg 1,4
}
\description{
eg 18 = 2 + 16 decomposes to 1, 4
}
\references{
yiming kang
\url{https://github.com/yiming-kang/rnaseq_pipe/blob/master/tools/utils.py}
}
