% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqObjectWithProtocolSpecificSizeFactors.R
\name{deseqObjectWithProtocolSpecificSizeFactors}
\alias{deseqObjectWithProtocolSpecificSizeFactors}
\title{Create a deseq data object with library protocol specific size factors}
\usage{
deseqObjectWithProtocolSpecificSizeFactors(passing_qc1_meta_qual, raw_counts)
}
\arguments{
\item{passing_qc1_meta_qual}{can be any metadata df, but if you're going to run deseq you may want to filter it for passing samples first}

\item{raw_counts}{a dataframe of raw counts with genes in the rows and samples in the columns. sample names must be the same as the fastqFileName
column in passing_qc1_meta_qual.}
}
\value{
a deseq data object with size factors calculated within the library protocol groups
}
\description{
Create a deseq data object with library protocol specific size factors
}
