% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{edgerLog2cpm}
\alias{edgerLog2cpm}
\title{Convenience function for edgeR's log2cpm}
\usage{
edgerLog2cpm(numeric_count_matrix, row_filter = NA)
}
\arguments{
\item{numeric_count_matrix}{output of \code{\link{countLibrary}} cast to data frame. See example}

\item{row_filter}{boolean vector length nrow(count_df) representing the rows to keep. Default is set to NA, which
returns all rows}
}
\value{
a numeric matrix in log2cpm
}
\description{
see \code{\link[edgeR]{cpm}}
}
\examples{

\dontrun{library(brentlabRnaSeqTools)
library(AnnotationDbi)
library(tidyverse)

cds_count = countLibrary(bam_path, kn99_db, 'cds', 'reverse', 8)

cds_counts_df = as_tibble(assay(cds_counts), rownames = "gene_name")

log2cpm = edgerLog2cpm(as.matrix(dplyr::select(cds_counts_df, -gene_name)))
log2cpm_df = as_tibble(log2cpm) \%>\%
  mutate(gene_name = cds_counts_df$gene_name)}

}
\seealso{
\code{\link[edgeR]{cpm}}
}
