% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{estimateSizeFactorsByProtocol}
\alias{estimateSizeFactorsByProtocol}
\alias{estimateSizeFactorsByProtocol,brentlabRnaSeqExperiment-method}
\title{Calculate Size Factors within protocol groups}
\usage{
estimateSizeFactorsByProtocol(x, ...)

\S4method{estimateSizeFactorsByProtocol}{brentlabRnaSeqExperiment}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqSet object}

\item{...}{other arguments}
}
\value{
a brentlabRnaSeqExperiment with sizeFactors estimated from within
protocol groups (SolexaPrep and not SolexaPrep)
}
\description{
Size Factors will be calculated for the set libraryProtocol 'SolexaPrep'
and the complement separately
}
