% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeseqFunctions.R
\name{examineSingleGroupWithLibDateSizeFactors}
\alias{examineSingleGroupWithLibDateSizeFactors}
\title{temporary function to examine only PBS samples, eg}
\usage{
examineSingleGroupWithLibDateSizeFactors(
  qc1_passing_metadata,
  raw_counts,
  row_filter,
  size_factor_subset_param
)
}
\arguments{
\item{row_filter}{is a boolean vector, eg qc1_passing_env_pert_meta_qual$MEDIUM == "PBS"}

\item{size_factor_subset_param}{eg LIBRARYDATE -- the column to group the samples for size factor calculation}
}
\value{
a list with slots metadata, raw_counts, size_factors
}
\description{
gets all samples in qc_passing_metadata with size_factor_subset_param same as samples in row_filter
}
