% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{exonRanges}
\alias{exonRanges}
\title{Given a GenomicFeatures annotation_db and a gene_id, extract an GRanges object of the exons}
\usage{
exon_ranges = exonRantes(kn99_db, "CKF44_05222")
}
\arguments{
\item{annotation_db}{a GenomicFeatures db. You can either get this from the bioconductor resources, or create your own with a gtf}

\item{gene_id}{the ID of a gene in the db. Eg, for cryptococcus CKF44_05222}
}
\value{
an IRanges object of the given gene's exons
}
\description{
Given a GenomicFeatures annotation_db and a gene_id, extract an GRanges object of the exons
}
\references{
GRanges, GRanges, GenomicFeatures
}
