% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{extractColData}
\alias{extractColData}
\alias{extractColData,brentlabRnaSeqSet-method}
\title{extract colData as tibble}
\usage{
extractColData(x, ...)

\S4method{extractColData}{brentlabRnaSeqSet}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqSet}

\item{...}{other arguments}
}
\value{
colData of the object as a tibble
}
\description{
extract colData as tibble
}
