% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{extractDesignMatrix}
\alias{extractDesignMatrix}
\alias{extractDesignMatrix,brentlabRnaSeqSet-method}
\title{extract the model.matrix using the design and colData}
\usage{
extractDesignMatrix(x, ...)

\S4method{extractDesignMatrix}{brentlabRnaSeqSet}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqSet}

\item{...}{other arguments}
}
\value{
the model matrix, using the design and coldata
}
\description{
extract the model.matrix using the design and colData
}
