% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{extractRLEByReplicateGroup_90minInduction}
\alias{extractRLEByReplicateGroup_90minInduction}
\title{These are messy and need to be cleaned up, but are how i calculate rle by replicate group for the 90minInduction and env_pert NOTE: only 90minInduction is currently tested with new database}
\usage{
extractRLEByReplicateGroup_90minInduction(
  meta_qual_df,
  norm_counts,
  output_dirpath,
  protocol_selector,
  already_logged_flag
)
}
\arguments{
\item{meta_qual_df}{is a dataframe pulled from the database with biosample --> quality assess (name is remnant of old system)}

\item{norm_counts}{from the deseq object}

\item{output_dirpath}{where to put the rle results. i suggest making a directory like "rle" in your working directory}

\item{protocol_selector}{whether or not to include library protocol in the replicate grouping}

\item{already_logged_flag}{are counts logged? (note: deseq norm counts are not logged unless you log them yourself)}
}
\value{
None. writes to file
}
\description{
These are messy and need to be cleaned up, but are how i calculate rle by replicate group for the 90minInduction and env_pert NOTE: only 90minInduction is currently tested with new database
}
