% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{filterWtByExperimentalLibdate_90min}
\alias{filterWtByExperimentalLibdate_90min}
\title{WT filtering helper for 90minInduction sets}
\usage{
filterWtByExperimentalLibdate_90min(blrs_90min)
}
\arguments{
\item{blrs_90min}{a brentlabRnaSeqExperiment object. this function is
written with the 90minInduction sets in mind.}
}
\description{
filter WT samples on libdates without perturbed samples
}
\note{
TODO: Move this to method and select function based on set_name
}
