% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMatrixFunctions.R
\name{fltrLowReplicateParams}
\alias{fltrLowReplicateParams}
\title{filter low replicate parameters from metadata}
\usage{
fltrLowReplicateParams(metadata_df, design_formula, replicate_threshold = 2)
}
\arguments{
\item{metadata_df}{a data frame that contains at least the model paramters of interest}

\item{design_formula}{an R formula, eg ~libraryDate+treatment, of parameters contained in the metadata_df}

\item{replicate_threshold}{the number of replicates below which samples will be removed. Default is 2}
}
\value{
the input metadata with samples in replicate groups with less than the specified thershold filtered out
}
\description{
given a model formula, remove samples with less than a specified number of replicates from the metadata
}
