% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{geneGRanges}
\alias{geneGRanges}
\title{Given a GenomicFeatures annotation_db and a gene_id, extract an GRanges
object of the cds}
\usage{
geneGRanges(annote_obj_path, gene_id, feature)
}
\arguments{
\item{annote_obj_path}{path to an annotation file parsed by rtracklayer::import}

\item{gene_id}{the ID of a gene in the db. Eg, for cryptococcus CKF44_05222}

\item{feature}{one of c("cds", "exon"), determines which feature to extract
from the annotations. Note that case is not important -- it will be handled
correctly in the function regardless.}
}
\value{
an IRanges object of the given gene's exons
}
\description{
Given a GenomicFeatures annotation_db and a gene_id, extract an GRanges
object of the cds
}
\references{
rtracklayer::import
}
