% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{getBamIndexPath}
\alias{getBamIndexPath}
\title{helper function to add .bai to bam path}
\usage{
getBamIndexPath(bamfile_path)
}
\arguments{
\item{bamfile_path}{path to bamfile}
}
\value{
bamfile_path concat with '.bai'
}
\description{
helper function to add .bai to bam path
}
