% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{getCapsuleImageSet}
\alias{getCapsuleImageSet}
\title{Retrieve, as a zip file, the set of images which correspond to a given
bioSampleNumber}
\usage{
getCapsuleImageSet(auth_token, bioSampleNumber, output_dir)
}
\arguments{
\item{auth_token}{see \code{\link{getUserAuthToken}}}

\item{bioSampleNumber}{The bioSampleNumber (from the bioSample table) for the
image}

\item{output_dir}{the directory on your local to which you wish to write
the zip file}
}
\value{
an httr request object
}
\description{
Retrieve, as a zip file, the set of images which correspond to a given
bioSampleNumber
}
\note{
if the request is successful, a zip file will be written to the output
dir
}
