% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Get Combined Metadata from the database}
\usage{
getMetadata(
  api_url,
  table_vector = c("BioSample", "RnaSample", "S1cdnaSample", "S2cdnaSample", "Library",
    "FastqFiles", "QualityAssess")
)
}
\arguments{
\item{api_url}{is the url to the api minus any table names, eg "http://13.59.167.2/api" No trailing /}

\item{table_vector}{list of tables to get from api_url. DEFAULT: c("BioSample", "RnaSample", "S1cdnaSample", "S2cdnaSample", "Library", "FastqFiles", "QualityAssess")}
}
\value{
a dataframe of the joined tables in the database
}
\description{
GET tables specified in table_vector, join on keys, return combined tables as a dataframe
}
