% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{getNfCoMultiqcDataPaths}
\alias{getNfCoMultiqcDataPaths}
\title{get relevant QC metrics tables from nf-co/rnaseq multiqc output}
\usage{
getNfCoMultiqcDataPaths(multiqc_output)
}
\arguments{
\item{multiqc_output}{path to multiqc_output}
}
\value{
list of full paths
}
\description{
get relevant QC metrics tables from nf-co/rnaseq multiqc output
}
