% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawCounts.R
\name{getRawCounts}
\alias{getRawCounts}
\title{get raw counts from database}
\usage{
getRawCounts(api_url, tablename = "Counts")
}
\arguments{
\item{api_url}{is the url to the api minus any table names, eg "http://13.59.167.2/api" No trailing /}
}
\value{
a dataframe of gene x sample
}
\description{
this downloads into $TEMPDIR a number of files (the paginations of the counts table) and then concats the counts.
}
\note{
it only takes the first 6967 genes (hard coded currently) which are the protein coding annotations in the current gff

column names are just the fastq file name (no .fastq.gz, no _read_count.tsv or any other suffix)
}
