% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{getUserAuthToken}
\alias{getUserAuthToken}
\title{get (via a http POST request) your user authentication token from the database}
\usage{
getUserAuthToken(url, username, password)
}
\arguments{
\item{url}{check the database_info variable. for configured organisms, you can find this under database_info$organism$token_auth}

\item{username}{a valid username for the database. If you don't have one, then you'll need to ask for one to be created}

\item{password}{password associated with your username}
}
\value{
the auth token associated with the username and password
}
\description{
get (via a http POST request) your user authentication token from the database
}
\note{
do not save your auth token in a public repository. For example, you might put it in your .Renviron and then make sure
that your .Renviron is in your .gitignore. Otherwise, save it outside of a github tracked directory or otherwise ensure
that it will not be pushed up to github
}
