% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{grant_df}
\alias{grant_df}
\title{the 2016 grant summary represented as a dataframe}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 164 rows and 3 columns.
}
\usage{
grant_df
}
\description{
also check the google sheet
}
\keyword{datasets}
