% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphYeastTimeCourse.R
\name{graphYeastTimeCourse}
\alias{graphYeastTimeCourse}
\title{Plot time vs normalized counts of a given gene over n samples, faceted by librarydate and run}
\usage{
graphYeastTimeCourse(metadata_df, genotype_1, gene_id, norm_counts)
}
\arguments{
\item{metadata_df}{where metadata$fastqFileName is equal in format to colnames(raw_counts)}

\item{genotype_1}{is the entry in the metadata genotype1 column which you would like to examine}

\item{gene_id}{is the gene id (the systematic name as opposed to the 'common name') most likely corresponding to genotype_1. This is how the correct row is extracted from the count data}

\item{norm_counts}{MUST have rownames assigned to the gene_ids includes at least all samples in metadata}
}
\value{
a ggplot graph of the gene_id in question, faceted on run and library date
}
\description{
graph_output = graphTimeCourse(cst6_sample_metadata, 'CST6', 'YIL036W', cst6_norm_counts)
}
\note{
see the vignette called yeast_timecourse_qc.Rmd. There is also sample data that comes with this
package, so you can run the vignette verbatim to see how it works. The vignette includes code that
}
