% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{htseq_libraryComplexity}
\alias{htseq_libraryComplexity}
\title{Calculate percentage of library accounted for by top expressed genes}
\usage{
htseq_libraryComplexity(htseq_filename, num_genes = 25)
}
\arguments{
\item{htseq_filename}{path to the htseq output for a given sample}

\item{num_genes}{used to select the top expressed n genes. Default is 25}
}
\value{
sum(top_n_genes) / total_counted_reads, rounded to 4 decimal places
}
\description{
Calculate percentage of library accounted for by top expressed genes
}
