% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{htseq_notAlignedTotalPercent}
\alias{htseq_notAlignedTotalPercent}
\title{Extract notAlignedTotalPercent from novoalign log}
\usage{
htseq_notAlignedTotalPercent(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=parseNovoalignLog]{parseNovoalignLog}}
  \describe{
    \item{\code{novoalign_log_path}}{path to a given sample's novoalign log}
  }}
}
\value{
unaligned reads / total reads
}
\description{
Extract notAlignedTotalPercent from novoalign log
}
