% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{htseq_proteinCodingTotal}
\alias{htseq_proteinCodingTotal}
\title{Calculate KN99 protein coding total}
\usage{
htseq_proteinCodingTotal(htseq_filename)
}
\arguments{
\item{htseq_filename}{path to the htseq output for a given sample}
}
\value{
The summed total of reads over protein coding genes (no nctr RNA)
}
\description{
Specifically for the 'old' pipeline with the 'old' annotations.
}
