% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{igv_script}
\alias{igv_script}
\title{create IGV viewer batch script (single range, as many tracks as you like)}
\usage{
igv_script(
  x,
  locus,
  output_dir = getwd(),
  image_basename = "",
  exit_browser = TRUE,
  maxPanelHeight = 500,
  ...
)
}
\arguments{
\item{locus}{the locus you wish to visualize. This will be used to filter
the granges attribute of the object}

\item{output_dir}{directory to which igv snapshot images will be deposited
by igv}

\item{image_basename}{basename of the igv snapshot files,
eg ckf44_12345_run2. The output_dir and additional locus details, etc
will be appended to this}

\item{exit_browser}{Default is TRUE, which means that the batchscript will be
written in a way that the browser will take an image, save it, and exit. Set
to false and this script may be used to launch an interactive session.}

\item{maxPanelHeight}{default 500. height of the IGV window}

\item{bam_list}{list of bamfiles -- multiple files means multiple tracks}
}
\value{
None. This prints a batch script to the output dir
}
\description{
this will create a batch script that may be run with the igv
browser. An example command is given in the examples below. See the extern
data for an example of what one of these scripts looks like.
}
\note{
run on the command line like so xvfb-run --auto-servernum igv.sh -b script.bat
}
