% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{is_integer64}
\alias{is_integer64}
\title{Test if value is datatype integer64}
\usage{
is_integer64(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
boolean with truth value determined by whether input is integer64
}
\description{
integer64 can cause some problems in filtering functions, eg
if timePoint is an integer64 datatype column, then
timePoint \%in\% c(0L, 1440L) and
timePoint == 0 | timePoint == 1440 return different subsets
}
\references{
\url{https://community.rstudio.com/t/how-to-handle-the-integer64-type/50024}
}
