% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kn99_novo_htseq_status}
\alias{kn99_novo_htseq_status}
\title{QC status codes for the novo+htseq pipeline}
\format{
An object of class \code{list} of length 8.
}
\usage{
kn99_novo_htseq_status
}
\description{
list of metrics and their statuses (powers of 2)
}
\keyword{datasets}
