% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kn99_novo_htseq_thresholds}
\alias{kn99_novo_htseq_thresholds}
\title{thresholds set for the novo+htseq pipeline}
\format{
An object of class \code{list} of length 10.
}
\usage{
kn99_novo_htseq_thresholds
}
\description{
list of metrics and their thresholds
}
\keyword{datasets}
