% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{listTables}
\alias{listTables}
\title{list tables in databse}
\usage{
listTables(db)
}
\arguments{
\item{db}{a connection to the database}
}
\value{
all tables in database
}
\description{
list tables in databse
}
\seealso{
\url{https://www.postgresqltutorial.com/postgresql-show-tables/}
}
