% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{loadAnnotationDatabase}
\alias{loadAnnotationDatabase}
\title{Read in annotationbi tx_db}
\usage{
loadAnnotationDatabase(annotation_db_path)
}
\arguments{
\item{annotation_db_path}{to .sqlite tx_db created with AnnotationDBI}
}
\value{
AnnotationDBI database obj
}
\description{
convenience function to read in an AnnotationDBI tx_db object created
with \link{}GenomicFeatures makeTxDbFromGFF
}
