% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{localView}
\alias{localView}
\title{To view a data.frame / data.table in LibreOffice Calc}
\usage{
localView(df)
}
\arguments{
\item{df}{the data.frame (or data.table or tibble)}
}
\description{
Copied from
\url{https://gitlab.com/zauster/ormisc/-/blob/master/R/view.R}.
The purpose of this function is to write the dataframe to at tmp file
and print the path. Use the clipr package to write the output to the
clipboard for easy pasting into a terminal. See example.
}
\examples{
\dontrun{
install.packages(clipr)
library(clipr)
# df is some dataframe in your environment
write_clip(localView(df))
# in a terminal, you can how paste in the line that the cmd above put
# in your clipboard and hit enter. your computer will know what to do next
# hopefully
}

}
\references{
\url{https://gitlab.com/zauster/ormisc/-/blob/master/R/view.R}
}
