% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{locusCoverage}
\alias{locusCoverage}
\title{Calculate coverage over a given feature}
\usage{
locusCoverage(
  bam_path,
  bam_index_path,
  annote_obj_path,
  gene_id,
  strandedness,
  feature = "cds",
  coverage_threshold = 0,
  ...
)
}
\arguments{
\item{bam_path}{path to a given samples alignment file (.bam)}

\item{bam_index_path}{path to a given sample's alignment file index (.bai)}

\item{annote_obj_path}{path to an annotation file parsed by rtracklayer::import}

\item{gene_id}{the ID of a gene in the db. Eg, for cryptococcus CKF44_05222}

\item{strandedness}{one of c("reverse", "same", "unstranded")}

\item{feature}{one of c("cds", "exon"), determines which feature to extract
from the annotations. Note that case is not important -- it will be handled
correctly in the function regardless.}

\item{coverage_threshold}{minimum threshold to consider. Default is 0}

\item{...}{other arguments to \link{strandedScanBamParam}}
}
\value{
coverage of feature
}
\description{
Calculate coverage over a given feature
}
