% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{locusLog2Cpm}
\alias{locusLog2Cpm}
\title{calculate log2cpm for a given locus}
\usage{
locusLog2Cpm(bam_path, strandedness, locus_granges, lib_size)
}
\arguments{
\item{bam_path}{path to bam file. Note: the index with extension .bai also must exist}

\item{strandedness}{the strandedness of the library}

\item{locus_granges}{a granges object specifying the locus over which to count}
}
\value{
either 0, if there are no counts, or the log2cpm of the counts over the locus. counting is done via the same
method as default HTSeq
}
\description{
calculate log2cpm for a given locus
}
