% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_pipeline.R
\name{moveNfCoFastqFiles}
\alias{moveNfCoFastqFiles}
\title{move fastq files from lts to scratch for nf-co/rnaseq pipeline}
\usage{
moveNfCoFastqFiles(nf_co_samplesheet, from_prefix)
}
\arguments{
\item{nf_co_samplesheet}{a samplesheet created by
createNfCorePipelineSamplesheet() with the prefixes set to what the path will
be once the files are moved to scratch}

\item{from_prefix}{filepath from source up to the filename of the file}
}
\value{
a dataframe with two columns, source and destination
}
\description{
using the samplesheet created by createNfCorePipelineSampleSheet,
create a file to move fastq files from lts to scratch
}
