% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{new_rlePlotFunc}
\alias{new_rlePlotFunc}
\title{plot RLE for a given column filter (eg, metadata\link{metadata$MEDIUM == 'PBS'}$FASTQFILENAME would give a list of fastqFileNames to filter)}
\usage{
new_rlePlotFunc(deseq_object, model_matrix, column_filter, title)
}
\arguments{
\item{deseq_object}{a deseq object with results from the DESeq() call}

\item{model_matrix}{the deseq_object model matrix}

\item{column_filter}{a vector of fastqFileNames (or whatever the columns -- samples -- are called)}

\item{title}{of the plots}
}
\value{
list with slots norm_count_rle and effect_removed_rle
}
\description{
plot RLE for a given column filter (eg, metadata\link{metadata$MEDIUM == 'PBS'}$FASTQFILENAME would give a list of fastqFileNames to filter)
}
