% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{outlierFence}
\alias{outlierFence}
\title{calculate lower/upper fence as a default threshold on qc metrics}
\usage{
outlierFence(metric_vector)
}
\arguments{
\item{metric_vector}{numeric vector on which to calculate lower/upper fence}
}
\value{
a list with the following slots: message, which stores a message
regarding the NA count, lower_inner, lower_outer, and upper_inner,
upper_outer which both store the fence values
}
\description{
calculate lower/upper inner and outer fences, and number of NAs
in metric vector. inner fence defined as Q1/Q3 -/+ 1.5\emph{IQR,
outer fence is the same but 3}IQR
}
