% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{parseBroadRnaseqcOutput}
\alias{parseBroadRnaseqcOutput}
\title{extract all metrics from broad rnaseqc package output}
\usage{
parseBroadRnaseqcOutput(rnaseqc_dir, bam_suffix = ".markdup.sorted.bam")
}
\arguments{
\item{rnaseqc_dir}{directory containing the rnaseqc output}

\item{bam_suffix}{the suffix to remove from the bam file sample names. Default to '.markdup.sorted.bam'
for nf-co/rnaseq_pipeline star_salmon output \url{https://nf-co.re/rnaseq}}
}
\value{
a list with items full_table, subset, rRna_vs_epe, rRna_vs_epe (see description)
}
\description{
extract, subset, reshape and plot metrics from the broad rnaseqc package
\url{https://github.com/getzlab/rnaseqc}. Function plots Expression Profiling Efficiency (epe)
vs rRNA percent and Estimated Library Complexity (elc) vs epe, both with marginal distributions.
}
\note{
this package is buggy (doesn't always get the metrics it says it will)
also, all metrics are included in the nf-co/pipeline

Estimated Library Complexity is a Picard metric and is similar to markDuplicates and totalDeduplicatedPercent
\url{https://gatk.broadinstitute.org/hc/en-us/articles/360037591931-EstimateLibraryComplexity-Picard-}
}
