% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{parseComparatives}
\alias{parseComparatives}
\title{parse text comparative sentence}
\usage{
parseComparatives(value1, comparative, value2)
}
\arguments{
\item{value1}{a number represented as a string, eg "3"}

\item{comparative}{one of c(">", ">=", "<", "<=", "==")}

\item{value2}{a number represented as a string, eg "4"}
}
\value{
a logical as if the characters were evaluated such as 3 > 4
}
\description{
return logical for sentence such as "3 > 4")
}
