% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{parseNovoalignLog}
\alias{parseNovoalignLog}
\title{Parse a novoalign log}
\usage{
parseNovoalignLog(novoalign_log_path)
}
\arguments{
\item{novoalign_log_path}{path to a given sample's novoalign log}
}
\value{
a dataframe of with columns metric, reads, percent. Metrics will
include: Read Sequences, Unique Alignment, Multi Mapped, No Mapping Found,
Homopolymer Filter
}
\description{
Parse a novoalign log
}
