% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{passing_genes_all}
\alias{passing_genes_all}
\title{genes considered NOT to have 'high dispersion'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6746 rows and 1 columns.
}
\usage{
passing_genes_all
}
\description{
union of old and new protocol
}
\keyword{datasets}
