% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{plotCoverageOverLocus}
\alias{plotCoverageOverLocus}
\title{plot coverage over locus}
\usage{
plotCoverageOverLocus(
  bamfile_path,
  annote_db,
  gene_id,
  strandedness,
  quality_threshold = 20L
)
}
\arguments{
\item{bamfile_path}{path to a bam file}

\item{annote_db}{a GenomicFeatures TxDb object. Maybe one made from a gtf, eg txdb = makeTxDbFromGFF("data/liftoff_h99_to_kn99.gtf", format = "gtf")}

\item{gene_id}{a gene_id of interest -- must be in the gene names of the annote_db object}

\item{strandedness}{one of c("reverse", "unstranded"). NOTE: forward only strand NOT currently configured}

\item{quality_threshold}{quality threshold above which reads will be considered. 20l is default, which is
chosen b/c it is the default for HTSeq}
}
\description{
ggbio plot with transcripts track and coverage track
}
