% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLogExpression.R
\name{plotRLEhistograms}
\alias{plotRLEhistograms}
\title{composite plot of all rle_stats in one plot}
\usage{
plotRLEhistograms(rle_df)
}
\arguments{
\item{rle_df}{a samples x rle stats df with at minimum columns SAMPLE_DEVIATION_MEDIAN, ABS_SAMPLE_DEVIATION_MEDIAN, INTERQUARTILE_RANGE}
}
\value{
a plot with three horizontal panels for each of the rle stats
}
\description{
composite plot of all rle_stats in one plot
}
\note{
no title on graphs. use the names of the returned object to title in presentation
}
