% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{postCapsuleJpeg}
\alias{postCapsuleJpeg}
\title{Send a single JPEG KN99 capsule image to the capsule image table}
\usage{
postCapsuleJpeg(auth_token, bioSampleNumber, jpeg_path)
}
\arguments{
\item{auth_token}{see \code{\link{getUserAuthToken}}}

\item{bioSampleNumber}{The bioSampleNumber (from the bioSample table) for the
image}

\item{jpeg_path}{Path on your local computer to the image}
}
\value{
an httr response object
}
\description{
Send a single JPEG KN99 capsule image to the capsule image table
}
