% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{postCapsuleJpeg_batch}
\alias{postCapsuleJpeg_batch}
\title{Send a set of JPEGs for a given bioSampleNumber to the database}
\usage{
postCapsuleJpeg_batch(
  auth_token,
  bioSampleNumber,
  image_dir,
  expected_number_of_images
)
}
\arguments{
\item{auth_token}{see \code{\link{getUserAuthToken}}}

\item{bioSampleNumber}{The bioSampleNumber (from the bioSample table) for the
image}

\item{image_dir}{Path on your local computer to the directory which stores
the images which relate to a single bioSampleNumber}

\item{expected_number_of_images}{The number of jpg files you expect to be
found in the image_dir}
}
\value{
a list of httr response objects
}
\description{
Send a set of JPEGs for a given bioSampleNumber to the database
}
