% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{postCounts}
\alias{postCounts}
\title{post counts to database}
\usage{
postCounts(
  database_counts_url,
  run_number,
  auth_token,
  new_counts_path,
  fastq_table,
  count_file_suffix = "_read_count.tsv"
)
}
\arguments{
\item{database_counts_url}{eg paste(database_info$kn99_base_url, "Counts/", sep="/")}

\item{run_number}{the run number of this counts sheet -- this is important b/c fastqFileNames aren't necessarily unique outside of their runs}

\item{auth_token}{see brentlabRnaSeqTools::getUserAuthToken()}

\item{new_counts_path}{path to the new counts csv}

\item{fastq_table}{a recent pull of the database fastq table}

\item{count_file_suffix}{the suffix appended to the fastqFileName in the count file column headings. default is "_read_count.tsv"}
}
\value{
a list of httr::response() objects
}
\description{
using the package httr, post the raw count .csv, which is the compiled counts for a given run, to the database
}
