% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{postFastqSheet}
\alias{postFastqSheet}
\title{post new fastq sheet to database}
\usage{
postFastqSheet(database_fastq_url, auth_token, new_fastq_path)
}
\arguments{
\item{database_fastq_url}{eg database_info$kn99_urls$FastqFiles. See see \code{\link{database_info}}}

\item{auth_token}{see \code{\link{getUserAuthToken}}}

\item{new_fastq_path}{path to new fastq sheet}
}
\description{
post new fastq sheet to database
}
