% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{postQcSheet}
\alias{postQcSheet}
\title{post new qc sheet to database}
\usage{
postQcSheet(database_qc_url, auth_token, run_number, new_qc_path, fastq_table)
}
\arguments{
\item{database_qc_url}{eg paste(database_info$kn99_base_url, "QualityAssessment/", sep="/")}

\item{auth_token}{see brentlabRnaSeqTools::getUserAuthToken}

\item{run_number}{the run number of this qc sheet -- this is important b/c fastqFileNames aren't necessarily unique
outside of their runs}

\item{new_qc_path}{path to the new counts csv}

\item{fastq_table}{a recent pull of the database fastq table}
}
\value{
a list of httr::response() objects
}
\description{
using the package httr, post the new qc sheet to the database
}
\note{
there can be problems with dependencies and the rename function. this is working for now,
but see here for more info \url{https://statisticsglobe.com/r-error-cant-rename-columns-that-dont-exist}
}
