% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{postTable}
\alias{postTable}
\title{Post a table to the database}
\usage{
postTable(database_table_url, auth_token, df)
}
\arguments{
\item{database_table_url}{see \code{\link{database_info}}. Use one of the URLS in the url slot}

\item{auth_token}{see \code{\link{getUserAuthToken}}}

\item{df}{a dataframe read in with, for example read_csv or vroom}
}
\value{
POST results object
}
\description{
Post a table to the database
}
