% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{proteinCodingCount}
\alias{proteinCodingCount}
\title{get total protein coding count from count dataframe}
\usage{
proteinCodingCount(counts, protein_coding_gene_ids)
}
\arguments{
\item{counts}{a dataframe with gene_ids in the rownames and (at minimum) a quantification column called raw_counts}
}
\description{
get total protein coding count from count dataframe
}
