% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{qaFilter}
\alias{qaFilter}
\alias{qaFilter,brentlabRnaSeqSet-method}
\alias{qaFilter,brentlabRnaSeqExperiment-method}
\title{Quality filter a brentlabRnaSeqSet object}
\usage{
qaFilter(x, ...)

\S4method{qaFilter}{brentlabRnaSeqSet}(x)

\S4method{qaFilter}{brentlabRnaSeqExperiment}(x, rle_iqr_threshold = NA, iqr_colname = NA)
}
\arguments{
\item{x}{a brentlabRnaSeqSet object}

\item{...}{other arguments}

\item{rle_iqr_threshold}{default is NA, which means no rle_iqr_threshold filtering.
Set a value to filter based on RLE IQR. Note: RLE stats are all calculated
after removing the libraryDate effect.}

\item{iqr_colname}{default is NA, which creates the iqr col name from the
set name in the object. Pass the col name itself, eg "my_set_iqr", to
specify a iqr column name manually}
}
\value{
a quality assessment filtered brentlabRnaSeqSet object

a quality assessment filtered brentlabRnaSeqExperiment object
}
\description{
Filter for manual passes (overrides auto fail),
automatic passes (unless auto failed), and optionally RLE IQR

Filter for manual passes (overrides auto fail),
automatic passes (unless auto failed), and optionally RLE IQR
}
\note{
on the brentlabRnaSeqSet (as opposed to the brentlabRnaSeqExperiment),
this only filters on manualAudit and autoAudit. no rle filtering

If there are less than 3 replicates, the RLE stats will be NA. These
are returned, also. Filter them out of the returned object if you wish to
have only those samples with more than 3 replicates.
}
