% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcExplorer.R
\name{qcExplorer}
\alias{qcExplorer}
\title{Explore QC metrics and set thresholds}
\usage{
qcExplorer(
  qc_database_path,
  sample_df_name,
  qc_metrics_df_name,
  id_field,
  batch_choices,
  debug = FALSE
)
}
\arguments{
\item{qc_database_path}{path to the sqlite database which contains the
the sample_df, qc_metrics_df, any QC metrics tables.
QC metrics tables must share an id column with each
other and the sample_df}

\item{sample_df_name}{name of the table which stores sample information, eg
id_column, libraryDate, condition}

\item{qc_metrics_df_name}{the name of the table which catalogs the metrics
which a user may threshold on, and the corresponding
comparative and 'status'}

\item{id_field}{the name of the common field between every table in the qc
database (except for the qc_metrics table, which is not a
table which is joined)}

\item{batch_choices}{a character vector, eg c("libraryDate", "libraryProtocol")
must correspond to columns in the sample_df. Controls
the options presented to the user for separating data
in the boxplot visualization tab}

\item{debug}{Logical. Default is false. Set to true to see more information
in the console. Intended for development debugging}
}
\description{
Interactively explore both user defined and nf-co/rnaseq_pipeline
QC metrics. Additionally, explore possible batch effects by
splitting data based on the sample_df input
}
