% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{qcGenotypeAndMarkerCoverage}
\alias{qcGenotypeAndMarkerCoverage}
\title{calcluate genotype and marker coverages for a given metadata df}
\usage{
qcGenotypeAndMarkerCoverage(metadata_df, annote_db, bam_prefix, bam_suffix)
}
\arguments{
\item{metadata_df}{a metadata sheet with at least the columns fastqFileName, genotype1, genotype2, runNumber, libraryProtocol}

\item{annote_db}{a connection to a local database -- see createQCdatabase()}

\item{bam_prefix}{the directory that contains the sequencing runs,, eg if locally mounted maybe "/mnt/htcf_lts/lts_align_expr"}

\item{bam_suffix}{the stuff appended to the end of the fastqFileName (minus the .fastq.gz). This might just be ".bam",
but could be something like "_sorted_aligned_reads_with_annote.bam"}
}
\value{
a sample sheet with the following columns: fastqFileName, fastqFileNumber, runNumber, column_param, locus, coverage
}
\description{
this produces a sheet that can be used to fill the 'custom' qc table for the brentlab kn99 and yeast databases
}
\note{
together, the bam_prefix/run_runNumber_samples/align/fastqFileName_bam_suffix form the path to the bamfile
}
