% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingCardsFunctions.R
\name{rank_response_plot}
\alias{rank_response_plot}
\title{Given expresion and binding signal, plot the rank response a la
doi:}
\usage{
rank_response_plot(
  expression_df_list,
  binding_df_list,
  tf,
  lfc_thres = 0,
  padj_thres = 0.05
)
}
\arguments{
\item{tf}{the name of the TF -- this will be the plot title}

\item{lfc_thres}{expression log2foldchange threshold. default is 0}

\item{padj_thres}{expression padj threshold. default is .05}

\item{expression_df}{a two column dataframe. One column must be
called 'gene' and must store gene identifiers (doesn't matter what, as long
as they are consistent both within the expression list and between the
expression list and the binding list). The second column identifies the
experiment and stores the expression value (eg, log2FC)}

\item{binding_df}{a list of dataframes with two columns. One column must be
called 'gene' and must correspond to the expression_df gene column. The
other column should be binding data. The column name should reference the
source, eg 'calling_cards' or 'chip_chip'.}
}
\value{
a ggplot object rank response plot
}
\description{
Given expresion and binding signal, plot the rank response a la
doi:
}
