% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{readInData}
\alias{readInData}
\title{read in columnar data}
\usage{
readInData(path)
}
\arguments{
\item{path}{path to a csv, tsv or xlsx}
}
\description{
given a csv, tsv or excel sheet, use the right function to read in the data
}
