% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLogExpression.R
\name{removeLibdateByReplicate}
\alias{removeLibdateByReplicate}
\title{Remove the libraryDate effects from KN99 dds objs}
\usage{
removeLibdateByReplicate(dds, replicate_colname)
}
\arguments{
\item{dds}{a deseq data object with either a formula or matrix in the design
note that if libraryDate is not included in the design, then this function
doesn't remove anything}

\item{replicate_colname}{which column to use as the replicate grouping. If
your replicates are described by more than 1 column, for this purpose,
group them to create 1 column. does not need to be the same as what is in
the design slot of the dds}
}
\value{
a list with both the unremoved and removed effect rle and summary
}
\description{
Remove the libraryDate effects from KN99 dds objs
}
