% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_and_tmp.R
\name{removeOneRedoIqr}
\alias{removeOneRedoIqr}
\title{progressively remove max IQR sample and recalculate}
\usage{
removeOneRedoIqr(sample_set, logged_norm_counts)
}
\description{
progressively remove max IQR sample and recalculate
}
