% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeParameterEffects.R
\name{removeParameterEffects}
\alias{removeParameterEffects}
\title{remove some effects from the counts}
\usage{
removeParameterEffects(
  deseq_object,
  design_matrix_of_effects_to_remove_with_intercept
)
}
\arguments{
\item{deseq_object}{a deseq data object}

\item{design_matrix_of_effects_to_remove:}{A possibly modified design matrix with any effects column that we do not want to remove set to 0}
}
\value{
a log2 scale gene by samples matrix with desired effects removed
}
\description{
subtract effect from norm counts of a single factor from coef x design. coef is in normalized log space
}
\note{
the design matrix will be transposed, and the intercept removed, in the function. Don't do that before
}
