% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{replicateByProtocolTally}
\alias{replicateByProtocolTally}
\alias{replicateByProtocolTally,brentlabRnaSeqExperiment-method}
\title{experimental replicates by libraryProtocol tally}
\usage{
replicateByProtocolTally(x, ...)

\S4method{replicateByProtocolTally}{brentlabRnaSeqExperiment}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqExperiment}

\item{...}{other arguments}
}
\description{
tally experimental replicate columns against the library prep
protocol
}
