% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSetFunctions.R
\name{replicateByProtocolTally_90min}
\alias{replicateByProtocolTally_90min}
\title{90min replicate by protocol tally}
\usage{
replicateByProtocolTally_90min(metadata_df)
}
\arguments{
\item{metadata_df}{sample metadata from the database or colData(blrs)}
}
\description{
tally 90min induction genotype1 replicates by libraryProtocol
}
