% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{rleByReplicateGroup}
\alias{rleByReplicateGroup}
\title{calculate RLE by replicate groups}
\usage{
rleByReplicateGroup(replicates_vector, gene_quants, log2_transformed_flag)
}
\arguments{
\item{replicates_vector}{a list of lists where each sublist represents a replicate group. Entries must be a metadata
parameter, such as fastqFileName, that corresponds to the columns of the counts.
Suggestion: use something like these dplyr functions to create the list of lists group_by() \%>\% group_split \%>\% pull(fastqFileName)}

\item{gene_quants}{a gene x sample dataframe with values as some sort of gene quantification (eg normed counts, or log2(norm_counts) with some effect removed), possibly already logged (@see already_logged_flag)}

\item{log2_transformed_flag}{a boolean where TRUE means the counts are already in log2 space}
}
\value{
a list of dataframes for each replicate group in replicateS_sample_list, each with dimensions gene x sample. values are RLE of the gene in a given sample
}
\description{
calculate RLE by replicate groups
}
\references{
rlePlotCompareEffectRemoved() to plot the norm counts and removedEffect 'counts' on the same plot
}
