% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{rlePlot_helper}
\alias{rlePlot_helper}
\title{the actual plotting function for rlePlot}
\usage{
rlePlot_helper(count_df, log2_transformed_flag, title)
}
\arguments{
\item{count_df}{counts in gene x sample}

\item{log2_transformed_flag}{boolean where TRUE indicates the counts are in log2 space}

\item{title}{title of the output plot}
}
\value{
a ggplot
}
\description{
the actual plotting function for rlePlot
}
