% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RleFunctions.R
\name{rlePlotCompareEffectRemoved}
\alias{rlePlotCompareEffectRemoved}
\title{plots output of rleSummaryByReplicateGroup}
\usage{
rlePlotCompareEffectRemoved(
  norm_counts_rle,
  removed_effect_rle,
  metadata_df,
  title
)
}
\arguments{
\item{norm_counts_rle}{output of calculateRLE (maybe one of the sublists in rleByReplicateGroup())}

\item{removed_effect_rle}{see norm_counts_rle, but after removing some batch effects}

\item{metadata_df}{metadata with at least FASTQFILENAME and LIBRARYDATE}

\item{title}{title of the plot}
}
\value{
a ggplot with both the norm counts (more transparent) and removedEffect 'counts' on the same plot
}
\description{
plots output of rleSummaryByReplicateGroup
}
