% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rleSummary.R
\name{rleSummary}
\alias{rleSummary}
\title{rleSummary calculates summary statistics of rleFullTable}
\usage{
rleSummary(rle_table_full)
}
\arguments{
\item{rle_table_full}{the output of rleFullTable}
}
\value{
a dataframe sample by rle summary statistics
}
\description{
rleSummary calculates summary statistics of rleFullTable
}
