% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLogExpression.R
\name{rleTransform}
\alias{rleTransform}
\title{transform normalized counts into regularized log expression}
\usage{
rleTransform(object)
}
\arguments{
\item{object}{a brentlabRnaSeqSet object}
}
\value{
a brentlabRnaSeqSetTransform object
}
\description{
transform normalized counts into regularized log expression
}
