% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{run_numbers_with_leading_zero}
\alias{run_numbers_with_leading_zero}
\title{A named list containing a run number without a leading zero, eg 647, with the value being the same runnumber with
a leading 0, eg 0647.}
\format{
An object of class \code{list} of length 13.
}
\usage{
run_numbers_with_leading_zero
}
\description{
this is best remedied in the database itself by forcing the column to be a string and adding the 0s
}
\keyword{datasets}
