% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectQcColumns.R
\name{selectQaColumns}
\alias{selectQaColumns}
\title{select fastqFileName, fastqFileNumber, and a pre-determined set of QC columns from a metadata df}
\usage{
selectQaColumns(metadata)
}
\arguments{
\item{metadata}{a metadata df with at least the columns listed in the select statement (see source code -- notably, must include interquartile range). Column names will be cast to uppper and returned as uppers}
}
\description{
select fastqFileName, fastqFileNumber, and a pre-determined set of QC columns from a metadata df
}
\note{
the column names for the metadata will be cast to upper and returned in upper

must include Interquartile range (think about removing this -- user could merge with IQR df after selecting these cols)
}
