% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{sendImageAnnotationsToDatabase}
\alias{sendImageAnnotationsToDatabase}
\title{Send Capsule Image Annotation Data to Database}
\usage{
sendImageAnnotationsToDatabase(
  annotation_path,
  auth_token,
  image_id,
  post_table = TRUE
)
}
\arguments{
\item{annotation_path}{path to the .m output of the matematica annotation script}

\item{auth_token}{your user authentication token to the database. See
\code{\link{getUserAuthToken}}. Note that this will be recorded as the person
responsible for these annotations -- please only upload annotations which
you did yourself}

\item{image_id}{numeric ID from the capsuleImage table which identifies a
given image}

\item{post_table}{boolean, default TRUE. Set to FALSE to prevent sending
the data to the database, and instead return the dataframe}
}
\value{
http response if post_table is true, otherwise the dataframe
}
\description{
Parse the .m output of the annotation script and send to the
database.
}
