% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetPerturbedLociToZero.R
\name{setPerturbedLociToZero}
\alias{setPerturbedLociToZero}
\title{Zero out perturbed expression}
\usage{
setPerturbedLociToZero(object)
}
\arguments{
\item{object}{a deseq data set column genotype1 in colData
(a brentlabRnaSeqSet also works)}
}
\description{
set expression of the gene in genotype1 for a particular
sample to zero
}
\note{
the error handling isn't good here yet -- it will error if there is a
perturbed loci that is not in the rownames of the object (eg, if it is filtered
due to low expression)
}
