% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpeff_summary_to_long.R
\name{snpeff_summary_to_long}
\alias{snpeff_summary_to_long}
\title{snpEff Summary to Long}
\usage{
snpeff_summary_to_long(
  snpeff_summary_path,
  sample_id,
  variant_caller,
  effect_threshold = 0,
  impact_threshold = 0
)
}
\arguments{
\item{snpeff_summary_path}{path to the .txt summary output of snpEff}

\item{sample_id}{unique identifier of this sample -- sample_id becomes a column}

\item{variant_caller}{the variant caller used to create the vcf}

\item{effect_threshold}{minimum count in a variant_effect factor level (see
the seealso link below for the snpEff docs if variant_effect factor doesn't
ring a bell)}

\item{impact_threshold}{minimum count in a variant_impact factor level (see
the command in effect_threshold)}
}
\value{
a tibble with the fields GeneId, TranscriptId, Biotype, effect,
effect_score, impact_score, impact, sample_id, variant_caller
}
\description{
Transform a snpEff tsv summary frame to long format
}
\examples{

\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\url{https://pcingola.github.io/SnpEff/se_outputsummary/}
}
