% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingCardsFunctions.R
\name{sort_rank_mean_expr}
\alias{sort_rank_mean_expr}
\title{Create a rank-response data frame}
\usage{
sort_rank_mean_expr(
  expression_df,
  expression_src,
  binding_df,
  binding_src,
  lfc_thres,
  padj_thres,
  rank_resolution = 5,
  num_rows = 30
)
}
\arguments{
\item{expression_df}{a three column dataframe with the column names
c('gene', 'log2FoldChange', 'padj') which describes expression data. It
may have more columns, but those three must exist}

\item{expression_src}{the source of the data, eg 'kemmeren' or 'brentlab'}

\item{binding_df}{a two column dataframe dataframe with the columns
c('gene', 'binding_signal'). It may have more columns, but those must exist}

\item{binding_src}{the source of the data, eg 'chip-chip' or 'calling-cards'}

\item{lfc_thres}{expression log2foldchange threshold. default is 0}

\item{padj_thres}{expression padj threshold. default is .05}

\item{rank_resolution}{how to bin ranks. Default is 10, which means that
ranks are binned into groups of 10 (eg the first group are the 10 highest)}

\item{num_rows}{the number of rows to return. Default is 15.
For instance, if the rank_resolution is 10, and you only want to look at
the first 150 ranked genes, then num_rows is 15.}
}
\value{
a rank_response dataframe
}
\description{
Create a rank-response data frame
}
