% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{splitProtocolGroups}
\alias{splitProtocolGroups}
\alias{splitProtocolGroups,brentlabRnaSeqSet-method}
\title{Split a brentlabRnaSeqSet into two by protocol}
\usage{
splitProtocolGroups(x, ...)

\S4method{splitProtocolGroups}{brentlabRnaSeqSet}(x)
}
\arguments{
\item{x}{a brentlabRnaSeqSet}

\item{...}{other arguments}
}
\value{
a list 'SolexaPrep' and 'E7420L' group. We call 'SolexaPrep' "old"
}
\description{
split an object into 'SolexaPrep' and 'E7420L'
}
