% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{class}
\name{VariantExplorer-class}
\alias{VariantExplorer-class}
\alias{create_variant_explorer}
\alias{igv_script,VariantExplorer-method}
\alias{summary,VariantExplorer-method}
\alias{visualize,VariantExplorer-method}
\title{VariantExplorer}
\usage{
create_variant_explorer(sample_metadata, gff_ranges, bsgenome, igv_genome = "")

\S4method{igv_script}{VariantExplorer}(
  x,
  locus,
  output_dir = getwd(),
  image_basename = "",
  exit_browser = TRUE,
  maxPanelHeight = 500,
  ...
)

\S4method{summary}{VariantExplorer}(object)

\S4method{visualize}{VariantExplorer}(x, gene, sample, plot_type = "pileup")
}
\arguments{
\item{sample_metadata}{sample metadata with at least the columns sample_id,
snpeff, variant_caller. Suggested columns are bam and vcf.}

\item{gff_ranges}{the gff for this organism imported with rtracklayer::import()}

\item{bsgenome}{path to a BSgenome object. KN99 has one -- check bioconductor}

\item{igv_genome}{path to a genome object created by igvtools}

\item{x}{a VariantExplorer object}

\item{object}{a VariantExplorer instance}

\item{gene}{the name of the gene you wish to visualize}

\item{sample}{the sample_id you wish to see}

\item{plot_type}{either 'pileup' or 'coverage'. Defaults to 'pileup'}
}
\value{
A VariantExplorer object
}
\description{
Use this to coordinate the analysis of the snpEff summary,
alignment files and bam files.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
\seealso{
\code{\link[Biostrings]{character(0)}}
\code{\link[GenomicRanges]{character(0)}}
}
