% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{archiveDatabase}
\alias{archiveDatabase}
\title{pull entire database (not counts) and save to output_dir for archival purposes}
\usage{
archiveDatabase(
  database_host,
  database_name,
  database_user,
  database_password,
  output_dir,
  archive_counts_flag = TRUE
)
}
\arguments{
\item{database_host}{if connecting to a database hosted on AWS, it might be something like ec2-54-83-201-96.compute-1.amazonaws.com}

\item{database_name}{name of the database, eg for cryptococcus kn99, the database might be named kn99_database. Check with the documentation, whoever set up the database, or get into the server and check directly}

\item{database_user}{a user of the actual database, with some level of permissions. You'll need to check with the database maintainer for this. It is suggested that you use a .Renviron file in your local project (make sure it is completely ignored by git, R, etc) to store this info}

\item{database_password}{password to the database user. You'll need to check with the database maintainer for this. It is suggested that you use a .Renviron file in your local project (make sure it is completely ignored by git, R, etc) to store this info}

\item{output_dir}{where to deposit a subdirectory, named by todays date in this format: 20210407, with the tables and combined_df inside. eg a mounted local directory /mnt/htcf_lts/crypto_database_archive/ --> /lts/mblab/Crypto/rnaseq_data/crypto_database_archive}

\item{archive_counts_flag}{boolean indicating whether or not to save the counts. default is TRUE}
}
\value{
None, writes a directory called <today's date> with tables and combined_df as .csv to output_dir
}
\description{
saves both the individual tables, including counts, and the combined_df
}
