% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveDatabase.R
\name{archiveDatabase}
\alias{archiveDatabase}
\title{pull entire database (not counts) and save to output_dir for archival purposes}
\usage{
archiveDatabase(output_dir)
}
\arguments{
\item{output_dir}{where to deposit a subdirectory called something like 20210302 (today's date)}
}
\value{
None, writes a directory called <today's date> with tables and combined_df as .csv to output_dir
}
\description{
saves both the individual tables and the combined_df
}
