% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{calculate coverage over a given locus}
\usage{
calculateCoverage(bamfile_path, annote_db, gene_id, strandedness, ...)
}
\arguments{
\item{bamfile_path}{path to a bam file @seealso brentlabRnaSeqTools::createBamPath()}

\item{annote_db}{a GenomicFeatures TxDb object. Maybe one made from a gtf, eg txdb = makeTxDbFromGFF("data/liftoff_h99_to_kn99.gtf", format = "gtf")}

\item{gene_id}{a gene_id of interest -- must be in the gene names of the annote_db object}

\item{...}{additional arguments to getCoverageOverRegion()}
}
\value{
percent coverage of feature with reads above a given quality threshold and coverage depth threshold (see getCoverageOverRegion())
}
\description{
calculate coverage over a given locus
}
