% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnvPertSet.R
\name{createEnvPertSet}
\alias{createEnvPertSet}
\title{filter combined_df for environmental perturbation sample set}
\usage{
createEnvPertSet(combined_df)
}
\arguments{
\item{combined_df}{the combined tables of the database}
}
\value{
environmental pertubation set
}
\description{
filter combined_df for environmental perturbation sample set
}
