% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgvFunctions.R
\name{createIgvBatchscript}
\alias{createIgvBatchscript}
\title{create IGV viewer batch script (single range, as many tracks as you like)}
\usage{
createIgvBatchscript(
  bam_list,
  granges,
  igv_genome,
  output_dir,
  output_file_basename,
  exit_browser = TRUE,
  maxPanelHeight = 500
)
}
\arguments{
\item{bam_list}{list of bamfiles -- multiple files means multiple tracks}

\item{granges}{granges describing the range you wish to visualize -- range
must be on one chromosome/contig}

\item{igv_genome}{.genome file created by IGV tools}

\item{output_dir}{where to deposit both the script and the browser shots}

\item{output_file_basename}{this will serve as both the name of the browser
shot after running IGV, and the name o the batchscript itself}

\item{exit_browser}{Default is TRUE, which means that the batchscript will be
written in a way that the browser will take an image, save it, and exit. Set
to false and this script may be used to launch an interactive session.}

\item{maxPanelHeight}{default 500. height of the IGV window}
}
\value{
None. This prints a batch script to the output dir
}
\description{
this will create a batch script that may be run with the igv
browser. An example command is given in the examples below. See the extern
data for an example of what one of these scripts looks like.
}
\note{
run on the command line like so xvfb-run --auto-servernum igv.sh -b script.bat
}
