% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{determineStrandedness}
\alias{determineStrandedness}
\title{from the metadata libraryProtocol column, determine the library strandedness.}
\usage{
determineStrandedness(library_protocol)
}
\arguments{
\item{library_protocol}{the library protocol of the sample (determines strandedness of the library)}
}
\value{
the strandedness of the library based on the value in the libraryProtocol column, or 'unstranded' by default
}
\description{
Currently set up for cryptococcus. E7420L returns reverse, SolexaPrep returns unstranded. default return is unstranded
}
\note{
: default is unstranded
}
