% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{getCoverageOverRegion}
\alias{getCoverageOverRegion}
\title{create a dataframe of coverage by nucleotide over a given locus}
\usage{
getCoverageOverRegion(
  bamfile_path,
  annote_db,
  gene_id,
  strandedness,
  quality_threshold = 20L,
  coverage_threshold = 0,
  lts_align_expr_prefix = Sys.getenv("LTS_ALIGN_EXPR_PREFIX"),
  bamfile_suffix = Sys.getenv("BAM_SUFFIX")
)
}
\arguments{
\item{bamfile_path}{path to a bam file @seealso brentlabRnaSeqTools::createBamPath()}

\item{annote_db}{a GenomicFeatures TxDb object. Maybe one made from a gtf, eg txdb = makeTxDbFromGFF("data/liftoff_h99_to_kn99.gtf", format = "gtf")}

\item{gene_id}{a gene_id of interest -- must be in the gene names of the annote_db object}

\item{strandedness}{one of c("reverse", "unstranded"). NOTE: forward only strand NOT currently configured}

\item{lts_align_expr_prefix}{= path to the directory which stores the run_12345_samples run directories.
For example, /lts/mblab/Crypto/rnaseq_data/lts_align_expr.
By default, this looks in your .Renviron for a key LTS_ALIGN_EXPR_PREFIX}

\item{bamfile_suffix}{= whatever is appended after the fastqFileName (no extension).
Currently, this is "_sorted_aligned_reads_with_annote.bam". By default, this looks in your
.Renviron for a key BAM_SUFFIX}
}
\description{
create a dataframe of coverage by nucleotide over a given locus
}
\references{
GRanges, RSamtools
}
