% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{getRawCounts}
\alias{getRawCounts}
\title{Get combined raw counts}
\usage{
getRawCounts(database_host, database_name, database_user, database_password)
}
\arguments{
\item{database_host}{if connecting to a database hosted on AWS,
it might be something like ec2-54-83-201-96.compute-1.amazonaws.com}

\item{database_name}{name of the database, eg for cryptococcus kn99, the database might be named kn99_database.
Check with the documentation, whoever set up the database, or get into the server and check
directly}

\item{database_user}{a user of the actual database, with some level of permissions. You'll need to check with the
database maintainer for this. It is suggested that you use a .Renviron file in your
local project (make sure it is completely ignored by git, R, etc) to store this info}

\item{database_password}{password to the database user. You'll need to check with the database maintainer for this.
It is suggested that you use a .Renviron file in your local project
(make sure it is completely ignored by git, R, etc) to store this info}
}
\value{
a gene by samples dataframe of all counts
}
\description{
Get combined raw counts
}
