% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawCounts.R
\name{getRawCounts}
\alias{getRawCounts}
\title{Get combined raw counts}
\usage{
getRawCounts(api_url)
}
\arguments{
\item{api_url}{NOTE: api_url is a variable saved into the project environment, in addition to the parameter. You can use the usage statement directly. The argument is provided for development in the event that you want to test a local instance of the database. An example url: "http://13.59.167.2/api" No trailing /}
}
\value{
a gene by samples dataframe of all counts
}
\description{
GET raw counts
}
