% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getRunNumberLeadingZero}
\alias{getRunNumberLeadingZero}
\title{correct run number to add leading zero where approprirate}
\usage{
getRunNumberLeadingZero(run_number)
}
\arguments{
\item{run_number}{a run number, most likely from the metadata runNumber field}
}
\description{
correct run number to add leading zero where approprirate
}
