% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunctions.R
\name{locusCoverage}
\alias{locusCoverage}
\title{Calculate coverage over a given feature}
\usage{
locusCoverage(
  bam_path,
  locus_granges,
  library_strandedness,
  index_suffix = ".bai",
  coverage_threshold = 0,
  ...
)
}
\arguments{
\item{bam_path}{path to a given samples alignment file (.bam)}

\item{locus_granges}{a granges object for a given gene
(or some other feature on only one strand)}

\item{library_strandedness}{one of c("reverse", "same", "unstranded")}

\item{index_suffix}{suffix to append to bampath. Default .bai}

\item{coverage_threshold}{minimum threshold to consider. Default is 0}

\item{...}{
  Arguments passed on to \code{\link[=strandedScanBamParam]{strandedScanBamParam}}
  \describe{
    \item{\code{quality_threshold}}{quality threshold above which reads will be
considered. 20l is default, which is chosen b/c it is the default for HTSeq}
  }}
}
\value{
coverage of feature
}
\description{
Calculate coverage over a given feature
}
