% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseInteractionFunctions.R
\name{postFastqSheet}
\alias{postFastqSheet}
\title{post new fastq sheet to database}
\usage{
postFastqSheet(database_fastq_url, auth_token, new_fastq_path)
}
\arguments{
\item{database_fastq_url}{eg database_info$kn99_urls$FastqFiles database_info is a saved data object in this package}

\item{auth_token}{see brentlabRnaSeqTools::getUserAuthToken()}

\item{new_fastq_path}{path to new fastq sheet}
}
\description{
post new fastq sheet to database
}
