% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityAssessmentFunctions.R
\name{qualityAssessmentFilter}
\alias{qualityAssessmentFilter}
\title{filter for manual passes (overrides auto fail) and automatic passes (unless auto failed)}
\usage{
qualityAssessmentFilter(metadata)
}
\arguments{
\item{metadata}{dataframe from the database}
}
\value{
a metadata dataframe with column names cast to upper
}
\description{
filter for manual passes (overrides auto fail) and automatic passes (unless auto failed)
}
