% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingCardsFunctions.R
\name{rank_response_plot}
\alias{rank_response_plot}
\title{Given expresion and binding signal, plot the rank response a la
doi:}
\usage{
rank_response_plot(
  expression_df_list,
  binding_df_list,
  tf,
  random_ratio,
  lfc_thres = 0,
  padj_thres = 0.05
)
}
\arguments{
\item{expression_df_list}{a NAMED list of gene expression dataframes. Each
dataframe must have the following columns, at minimum:
c('gene', "log2FoldChange", "padj"). Names of items in the list must
correspond to the data source (eg, 'kemmeren', 'brentlab', etc)}

\item{binding_df_list}{a NAMED list of gene expression dataframes. Each
dataframe must have the following columns, at minimum:
c('gene', 'binding_signal'). Names of items in the list must
correspond to the data source (eg, 'chip-chip', 'calling-cards', etc)}

\item{tf}{the name of the TF -- this will be the plot title}

\item{lfc_thres}{expression log2foldchange threshold. default is 0}

\item{padj_thres}{expression padj threshold. default is .05}
}
\value{
a ggplot object rank response plot
}
\description{
Given expresion and binding signal, plot the rank response a la
doi:
}
