% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingPipelineFunction.R
\name{readHTSeqFile}
\alias{readHTSeqFile}
\title{Parse HTSeq output to tibble}
\usage{
readHTSeqFile(fname, samplename = "sample")
}
\arguments{
\item{fname}{filepath to htseq output}

\item{samplename}{in the legacy pipeline, the fastqFileName with .fastq.gz
replaced with _read_count.tsv}
}
\value{
a dataframe with columns feature, 'sample'
}
\description{
parse the htseq output as a tibble with columns (feature, sample).
This code is copied from the National Cancer Institute GenomicDataCommons
}
\references{
\href{https://bioconductor.org/packages/release/bioc/html/GenomicDataCommons.html}{GenomicDataCommons::readHTSeqFile}
}
