% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLogExpression.R
\name{removeLibdateByReplicate}
\alias{removeLibdateByReplicate}
\title{composite plot of all rle_stats in one plot}
\usage{
removeLibdateByReplicate(dds, replicate_colname)
}
\arguments{
\item{dds}{a deseq data object with either a formula or matrix in the design
note that if libraryDate is not included in the design, then this function
doesn't remove anything}

\item{replicate_colname}{which column to use as the replicate grouping. If
your replicates are described by more than 1 column, for this purpose,
group them to create 1 column. does not need to be the same as what is in
the design slot of the dds}

\item{rle_df}{a samples x rle stats df with at minimum columns SAMPLE_DEVIATION_MEDIAN, ABS_SAMPLE_DEVIATION_MEDIAN, INTERQUARTILE_RANGE}
}
\value{
a plot with three horizontal panels for each of the rle stats

Remove the libraryDate effects from KN99 dds objs

a list with both the unremoved and removed effect rle and summary
}
\description{
composite plot of all rle_stats in one plot
}
\note{
no title on graphs. use the names of the returned object to title in presentation
}
