% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_variables.R
\name{runSVA}
\alias{runSVA}
\title{run SVA}
\usage{
runSVA(metadata, raw_counts, known_covariates_formula)
}
\arguments{
\item{metadata}{metadata in the shape samples x parameters where the columns are features of the data (eg, libraryDate, RIN, etc)}

\item{raw_counts}{raw gene counts in the shape gene x samples where ncols matches nrow of metadata (samples == samples)}

\item{known_covariate_formula}{eg ~Condition+Cell, passed as an R formula}
}
\description{
run SVA
}
