% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_variables.R
\name{runSVA}
\alias{runSVA}
\title{run SVA}
\usage{
runSVA(raw_counts, null_model_matrix, full_model_matrix)
}
\arguments{
\item{raw_counts}{raw gene counts in the shape gene x samples where ncols matches nrow of metadata (samples == samples)}

\item{null_model_matrix}{a model matrix respresenting only the batch effects. Could possibly be intercept only}

\item{full_model_matrix}{the full model describing the experiment. Critically, this includes the parameter of interest}
}
\description{
run SVA
}
