% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{sendCountsToDatabase}
\alias{sendCountsToDatabase}
\title{Send Counts to Database (legacy pipeline)}
\usage{
sendCountsToDatabase(htseq_path, run_number, auth_token, fastq_table_path)
}
\arguments{
\item{htseq_path}{path to htseq-counts output}

\item{run_number}{run number from which the sample originates}

\item{auth_token}{your user authentication token to the database. See
\code{\link{getUserAuthToken}}}

\item{fastq_table_path}{path to recent fastq table reflecting current state of
database}
}
\value{
http response
}
\description{
for use in the novoalign + htseq brentlab rnaseq pipeline.
}
