% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingCardsFunctions.R
\name{sort_rank_mean_expr}
\alias{sort_rank_mean_expr}
\title{Create a rank-response data frame}
\usage{
sort_rank_mean_expr(
  expression_df,
  expression_src,
  binding_df,
  binding_src,
  lfc_thres,
  padj_thres,
  rank_resolution = 10,
  num_rows = 15
)
}
\arguments{
\item{expression_df}{a two column dataframe. One column must be
called 'gene' and must store gene identifiers (doesn't matter what, as long
as they are consistent both within the expression list and between the
expression list and the binding list). The second column identifies the
experiment and stores the expression value (eg, log2FC)}

\item{expression_src}{the source of the data, eg 'kemmeren' or 'brentlab'}

\item{binding_df}{a list of dataframes with two columns. One column must be
called 'gene' and must correspond to the expression_df gene column. The
other column should be binding data. The column name should reference the
source, eg 'calling_cards' or 'chip_chip'.}

\item{binding_src}{the source of the data, eg 'chip-chip' or 'calling-cards'}

\item{lfc_thres}{expression log2foldchange threshold. default is 0}

\item{padj_thres}{expression padj threshold. default is .05}

\item{rank_resolution}{how to bin ranks. Default is 10, which means that
ranks are binned into groups of 10 (eg the first group are the 10 highest)}

\item{num_rows}{the number of rows to return. Default is 15.
For instance, if the rank_resolution is 10, and you only want to look at
the first 150 ranked genes, then num_rows is 15.}
}
\value{
a rank_response dataframe
}
\description{
Create a rank-response data frame
}
