% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFunctions.R
\name{strandedScanBamParam}
\alias{strandedScanBamParam}
\title{create coverage scanbamparam object}
\usage{
strandedScanBamParam(locus_granges, strandedness, quality_threshold = 20L)
}
\arguments{
\item{locus_granges}{a granges object for a given gene (or some other feature on only one strand)}

\item{strandedness}{one of c("reverse", "same", "unstranded")}

\item{quality_threshold}{quality threshold above which reads will be considered. 20l is default, which is
chosen b/c it is the default for HTSeq}
}
\value{
a ScanBamParam object with certain configured options, as well as some reasonable defaults, to filter
a bam file for reads of interest based on the strandedness (protocol) of the library prep.
}
\description{
helper function to create ScanBamParam object with appropriate strandedness information
}
\seealso{
\code{\link[Rsamtools]{ScanBamParam}}
}
